Move helper methods our of price registry
