+++ core/services/ocr2/plugins/ccip/internal/ccipdata/v1_2_0/price_registry.go
@@ -3,11 +3,9 @@ package v1_2_0
 import (
 	"context"
 	"math/big"
-	"testing"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/stretchr/testify/require"
 
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/client"
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/logpoller"
@@ -60,34 +58,4 @@ func (p *PriceRegistry) GetTokenPrices(ctx context.Context, wantedTokens []commo
 		})
 	}
 	return tpu, nil
-}
-
-// ApplyPriceRegistryUpdate is a helper function used in tests only.
-func ApplyPriceRegistryUpdate(t *testing.T, user *bind.TransactOpts, addr common.Address, ec client.Client, gasPrices []ccipdata.GasPrice, tokenPrices []ccipdata.TokenPrice) common.Hash {
-	require.True(t, len(gasPrices) <= 1)
-	pr, err := price_registry.NewPriceRegistry(addr, ec)
-	require.NoError(t, err)
-	o, err := pr.Owner(nil)
-	require.NoError(t, err)
-	require.Equal(t, user.From, o)
-	var tps []price_registry.InternalTokenPriceUpdate
-	for _, tp := range tokenPrices {
-		tps = append(tps, price_registry.InternalTokenPriceUpdate{
-			SourceToken: tp.Token,
-			UsdPerToken: tp.Value,
-		})
-	}
-	var gps []price_registry.InternalGasPriceUpdate
-	for _, gp := range gasPrices {
-		gps = append(gps, price_registry.InternalGasPriceUpdate{
-			DestChainSelector: gp.DestChainSelector,
-			UsdPerUnitGas:     gp.Value,
-		})
-	}
-	tx, err := pr.UpdatePrices(user, price_registry.InternalPriceUpdates{
-		TokenPriceUpdates: tps,
-		GasPriceUpdates:   gps,
-	})
-	require.NoError(t, err)
-	return tx.Hash()
 }
